+++ noodles-sam/examples/sam_read_header.rs
@@ -10,7 +10,6 @@ fn main() -> io::Result<()> {
     let src = env::args().nth(1).expect("missing src");
 
     let mut reader = sam::reader::Builder::default().build_from_path(src)?;
-
     let header = reader.read_header()?;
     print!("{header}");