+++ noodles-sam/examples/sam_reheader.rs
@@ -8,11 +8,11 @@ use std::{env, io};
 
 use noodles_sam as sam;
 
-fn main() -> Result<(), Box<dyn std::error::Error>> {
+fn main() -> io::Result<()> {
     let src = env::args().nth(1).expect("missing src");
 
     let mut reader = sam::reader::Builder::default().build_from_path(src)?;
-    let mut header: sam::Header = reader.read_header()?.parse()?;
+    let mut header = reader.read_header()?;
 
     header.add_comment("a comment added by noodles-sam");