+++ noodles-sam/src/reader/header.rs
@@ -1,6 +1,18 @@
 use std::io::{self, BufRead};
 
-pub(super) fn read_header<R>(reader: &mut R) -> io::Result<String>
+use crate::Header;
+
+pub(super) fn read_header<R>(reader: &mut R) -> io::Result<Header>
+where
+    R: BufRead,
+{
+    read_raw_header(reader).and_then(|s| {
+        s.parse()
+            .map_err(|e| io::Error::new(io::ErrorKind::InvalidData, e))
+    })
+}
+
+fn read_raw_header<R>(reader: &mut R) -> io::Result<String>
 where
     R: BufRead,
 {
@@ -40,29 +52,29 @@ mod tests {
     use super::*;
 
     #[test]
-    fn test_read_header_with_no_header() -> io::Result<()> {
+    fn test_read_raw_header_with_no_header() -> io::Result<()> {
         let data = b"*\t4\t*\t0\t255\t*\t*\t0\t0\t*\t*\n";
         let mut reader = &data[..];
-        assert!(read_header(&mut reader)?.is_empty());
+        assert!(read_raw_header(&mut reader)?.is_empty());
         Ok(())
     }
 
     #[test]
-    fn test_read_header_with_no_records() -> io::Result<()> {
+    fn test_read_raw_header_with_no_records() -> io::Result<()> {
         let data = "@HD\tVN:1.6\n";
         let mut reader = data.as_bytes();
-        let header = read_header(&mut reader)?;
+        let header = read_raw_header(&mut reader)?;
         assert_eq!(header, data);
         Ok(())
     }
 
     #[test]
-    fn test_read_header_with_multiple_buffer_fills() -> io::Result<()> {
+    fn test_read_raw_header_with_multiple_buffer_fills() -> io::Result<()> {
         use std::io::BufReader;
 
         let data = "@HD\tVN:1.6\n@SQ\tSN:sq0\tLN:8\n";
         let mut reader = BufReader::with_capacity(16, data.as_bytes());
-        let header = read_header(&mut reader)?;
+        let header = read_raw_header(&mut reader)?;
 
         assert_eq!(header, data);